Refactor onit cluster controller to separate each layer of resources into a separate go file.
