+++ test/runner/client.go
@@ -1,0 +1,96 @@
+// Copyright 2019-present Open Networking Foundation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package runner
+
+import (
+	"errors"
+	atomixk8s "github.com/atomix/atomix-k8s-controller/pkg/client/clientset/versioned"
+	apiextension "k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/tools/clientcmd"
+	"os"
+	"path/filepath"
+)
+
+// newKubeClient returns a new Kubernetes client from the environment
+func newKubeClient() (*kubernetes.Clientset, error) {
+	kubeconfig := os.Getenv("KUBECONFIG")
+	if kubeconfig == "" {
+		home := homeDir()
+		if home == "" {
+			return nil, errors.New("no home directory configured")
+		}
+		kubeconfig = filepath.Join(home, ".kube", "config")
+	}
+
+	// use the current context in kubeconfig
+	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	// create the clientset
+	return kubernetes.NewForConfig(config)
+}
+
+// newExtensionsKubeClient returns a new extensions API server Kubernetes client from the environment
+func newExtensionsKubeClient() (*apiextension.Clientset, error) {
+	kubeconfig := os.Getenv("KUBECONFIG")
+	if kubeconfig == "" {
+		home := homeDir()
+		if home == "" {
+			return nil, errors.New("no home directory configured")
+		}
+		kubeconfig = filepath.Join(home, ".kube", "config")
+	}
+
+	// use the current context in kubeconfig
+	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	// create the clientset
+	return apiextension.NewForConfig(config)
+}
+
+// newAtomixKubeClient returns a new Atomix Kubernetes client from the environment
+func newAtomixKubeClient() (*atomixk8s.Clientset, error) {
+	kubeconfig := os.Getenv("KUBECONFIG")
+	if kubeconfig == "" {
+		home := homeDir()
+		if home == "" {
+			return nil, errors.New("no home directory configured")
+		}
+		kubeconfig = filepath.Join(home, ".kube", "config")
+	}
+
+	// use the current context in kubeconfig
+	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	// create the clientset
+	return atomixk8s.NewForConfig(config)
+}
+
+// homeDir returns the user's home directory if defined by environment variables
+func homeDir() string {
+	if h := os.Getenv("HOME"); h != "" {
+		return h
+	}
+	return os.Getenv("USERPROFILE") // windows
+}
